requirements.txt
streamlit
google-generativeai
crewai
crewai-tools
